Update pixel-display V10 interface
* Add blockingRoi in the HistogramConfig.
For the histogram client uses V10 or later interface, it can either skip
the blockingRoi field or set with (0, 0, 0, 0) to disable blocking ROI.
Otherwise, the blocking ROI will be enabled.
For the histogram client uses V9 interface, blockingRoi will be null by
default.

* HistogramController will implement the disable ROI.
In order to capture the histogram of the entire screen, histogram client
can either specify the ROI with screen coordinates or set ROI with
(0, 0, 0, 0) to disable the ROI.

Bug: 299410182
Test: Build pass
Change-Id: I5ef20244163d48811d8789539f2d72f9697f05c2
Signed-off-by: Leo Chen <yinchiuan@google.com>